CMake + opengl

+++ Widgets/src/geom/Dimension.h
@@ -1,6 +1,6 @@
 #pragma once
 
 struct Dimension {
-	int width, height;
-	Dimension(int w, int h);
+	float width, height;
+	Dimension(float w, float h);
 };
