+++ Widgets/src/geom/Position.h
@@ -1,6 +1,6 @@
 #pragma once
 
 struct Position {
-	int x, y;
-	Position(int _x, int _y);
+	float x, y;
+	Position(float _x, float _y);
 };
